final practice for week

+++ phone-store-express/routes/phone-api-router.js
@@ -55,6 +55,10 @@ router.post("/phones", (req, res, next) => {
 
 
 router.get("/phones/:id", (req, res, next) => {
+  if (req.user === undefined) {
+    res.status(400).json({ error: "Not logged in" });
+    return;
+  }
   Phone.findById(req.params.id)
   .then((phoneFromDb) => {
     if (phoneFromDb === null) {
@@ -78,6 +82,10 @@ router.get("/phones/:id", (req, res, next) => {
 
 
 router.delete("/phones/:id", (req, res, next) => {
+  if (req.user === undefined) {
+    res.status(400).json({ error: "Not logged in" });
+    return;
+  }
   Phone.findByIdAndRemove(req.params.id)
   .then((phoneFromDb) => {
     if (phoneFromDb === null) {
